scraper/requirements.txt
beautifulsoup4==4.11.2
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
requests==2.28.1
